Caching system

+++ src/main/java/me/artuto/endless/handlers/StarboardHandler.java
@@ -18,10 +18,11 @@
 package me.artuto.endless.handlers;
 
 import me.artuto.endless.Bot;
-import me.artuto.endless.data.managers.GuildSettingsDataManager;
-import me.artuto.endless.data.managers.StarboardDataManager;
-import me.artuto.endless.entities.StarboardMessage;
+import me.artuto.endless.storage.data.managers.GuildSettingsDataManager;
+import me.artuto.endless.storage.data.managers.StarboardDataManager;
+import me.artuto.endless.core.entities.StarboardMessage;
 import me.artuto.endless.utils.FinderUtil;
+import me.artuto.endless.utils.GuildUtils;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.MessageBuilder;
 import net.dv8tion.jda.core.entities.*;
@@ -62,7 +63,7 @@ public class StarboardHandler
 
             if(!(isConfigured(guild))) return;
 
-            TextChannel starboard = gsdm.getStarboardChannel(guild);
+            TextChannel starboard = GuildUtils.getStarboardChannel(event.getGuild());
             Message starredMsg = getMessage(event.getMessageIdLong(), event.getChannel());
             if(starredMsg==null)
                 return;
@@ -123,7 +124,7 @@ public class StarboardHandler
                 return;
 
             StarboardMessage starboardMsg = sdm.getStarboardMessage(starredMsg.getIdLong());
-            TextChannel starboard = gsdm.getStarboardChannel(event.getGuild());
+            TextChannel starboard = GuildUtils.getStarboardChannel(event.getGuild());
 
             if(existsOnStarboard(starredMsg.getIdLong()))
             {
@@ -158,12 +159,12 @@ public class StarboardHandler
 
     private static boolean isConfigured(Guild guild)
     {
-        return !(gsdm.getStarboardChannel(guild) == null) && !(gsdm.getStarboardCount(guild) == null);
+        return !(GuildUtils.getStarboardChannel(guild) == null) && !(GuildUtils.getStarboardCount(guild) == 0);
     }
 
     private static boolean amountPassed(Message msg)
     {
-        return getStarCount(msg) >= gsdm.getStarboardCount(msg.getGuild());
+        return getStarCount(msg) >= GuildUtils.getStarboardCount(msg.getGuild());
     }
 
     private static int getStarCount(Message msg)
@@ -184,7 +185,7 @@ public class StarboardHandler
 
     private static void updateCount(Message msg, Long starboardMsg, Integer amount)
     {
-        TextChannel tc = gsdm.getStarboardChannel(msg.getGuild());
+        TextChannel tc = GuildUtils.getStarboardChannel(msg.getGuild());
         tc.getMessageById(starboardMsg).queue(s -> s.editMessage(getEmote(amount)+" **"+amount+"** "+msg.getTextChannel().getAsMention()+" ID: "+msg.getId()).queue(null, null), null);
     }
 
@@ -203,6 +204,9 @@ public class StarboardHandler
 
     private static void delete(TextChannel starboard, StarboardMessage starboardMsg)
     {
+        if(starboard==null)
+            return;
+
         starboard.getMessageById(starboardMsg.getStarboardMessageId()).queue(s -> {
             s.delete().queue();
             sdm.deleteMessage(starboardMsg.getMessageIdLong(), starboardMsg.getStarboardMessageIdLong());
@@ -211,7 +215,7 @@ public class StarboardHandler
 
     private static void check(Guild guild, long msg)
     {
-        TextChannel starboard = gsdm.getStarboardChannel(guild);
+        TextChannel starboard = GuildUtils.getStarboardChannel(guild);
         StarboardMessage starboardMsg = sdm.getStarboardMessage(msg);
 
         if(existsOnStarboard(msg))
@@ -226,7 +230,7 @@ public class StarboardHandler
         }
         catch(ErrorResponseException e)
         {
-            delete(gsdm.getStarboardChannel(tc.getGuild()), sdm.getStarboardMessage(id));
+            delete(GuildUtils.getStarboardChannel(tc.getGuild()), sdm.getStarboardMessage(id));
             return null;
         }
     }
